+++ vite-project/src/App.jsx
@@ -2,21 +2,21 @@ import { useEffect, useState } from 'react'
 
 import './App.css'
 
-import PopExit from './components/popups/PopExit.jsx'
-import PopNewCard from './components/popups/PopNewCard.jsx'
-import PopBrowse from './components/popups/PopBrowse.jsx'
+import PopExit from './components/popups/PopExit/PopExit.jsx'
+import PopNewCard from './components/popups/PopNewCard/PopNewCard.jsx'
+import PopBrowse from './components/popups/PopBrowse/PopBrowse.jsx'
 import Header from './components/Header/Header.jsx'
 import Main from './components/Main/Main.jsx'
 import { Wrapper } from "./App.styled.js";
-import { GlobalStyle } from "./Global.styled";
+import { GlobalStyle } from "./Global.styled.js";
 
 
 
 function App() {
 
-	useEffect(()=> {
-		setTimeout(()=>{
-		 setIsLoaded(false);
+	useEffect(() => {
+		setTimeout(() => {
+			setIsLoaded(false);
 		}, 2000)
 	}, [])
 
@@ -45,24 +45,24 @@ function App() {
 				id: self.crypto.randomUUID(),
 				theme: "Web Design",
 				title: "Название задачи",
-        		date: "30.10.23",
-        		status: "Без статуса",
+				date: "30.10.23",
+				status: "Без статуса",
 
 			},
 			{
 				id: self.crypto.randomUUID(),
 				theme: "Copywriting",
 				title: "Название задачи",
-        		date: "30.10.23",
-        		status: "Без статуса",
+				date: "30.10.23",
+				status: "Без статуса",
 
 			},
 			{
 				id: self.crypto.randomUUID(),
 				theme: "Web Design",
 				title: "Название задачи",
-        		date: "30.10.23",
-        		status: "Без статуса",
+				date: "30.10.23",
+				status: "Без статуса",
 
 			},
 		],
@@ -71,8 +71,8 @@ function App() {
 				id: self.crypto.randomUUID(),
 				theme: "Research",
 				title: "Название задачи",
-        		date: "30.10.23",
-        		status: "Без статуса",
+				date: "30.10.23",
+				status: "Без статуса",
 
 			},
 		],
@@ -81,24 +81,24 @@ function App() {
 				id: self.crypto.randomUUID(),
 				theme: "Research",
 				title: "Название задачи",
-        		date: "30.10.23",
-        		status: "Без статуса",
+				date: "30.10.23",
+				status: "Без статуса",
 
 			},
 			{
 				id: self.crypto.randomUUID(),
 				theme: "Copywriting",
 				title: "Название задачи",
-        		date: "30.10.23",
-        		status: "Без статуса",
+				date: "30.10.23",
+				status: "Без статуса",
 
 			},
 			{
 				id: self.crypto.randomUUID(),
 				theme: "Web Design",
 				title: "Название задачи",
-        		date: "30.10.23",
-        		status: "Без статуса",
+				date: "30.10.23",
+				status: "Без статуса",
 
 			},
 		],
@@ -107,8 +107,8 @@ function App() {
 				id: self.crypto.randomUUID(),
 				theme: "Research",
 				title: "Название задачи",
-        		date: "30.10.23",
-        		status: "Без статуса",
+				date: "30.10.23",
+				status: "Без статуса",
 
 			},
 		],
@@ -117,8 +117,8 @@ function App() {
 				id: self.crypto.randomUUID(),
 				theme: "Research",
 				title: "Название задачи",
-        		date: "30.10.23",
-        		status: "Без статуса",
+				date: "30.10.23",
+				status: "Без статуса",
 
 			},
 		]
@@ -131,7 +131,7 @@ function App() {
 	return (
 		<>
 
-		    <GlobalStyle />
+			<GlobalStyle />
 
 			<Wrapper >
 				<PopExit />
@@ -140,7 +140,7 @@ function App() {
 				<Header />
 				<Main cards={cards} isLoaded={isLoaded} />
 			</Wrapper >
-			
+
 		</>
 
 	)
