+++ vite-project/src/components/popups/PopBrowse/PopBrowse.jsx
@@ -1,8 +1,8 @@
-import CalendarBox from "../Calendar/Calendar.jsx";
+import CalendarBox from "../../Calendar/Calendar.jsx";
 
 function PopBrowse() {
-    return (
-        <div className="pop-browse" id="popBrowse">
+	return (
+		<div className="pop-browse" id="popBrowse">
 			<div className="pop-browse__container">
 				<div className="pop-browse__block">
 					<div className="pop-browse__content">
@@ -66,6 +66,6 @@ function PopBrowse() {
 				</div>
 			</div>
 		</div>
-    )
+	)
 }
 export default PopBrowse
