+++ vite-project/src/lib/appRoutes.js
@@ -1,0 +1,8 @@
+export const appRoutes = {
+  MAIN: "/",
+  TASK: "/task/:id",
+  SIGNIN: "/signin",
+  SIGNUP: "/signup",
+  EXIT: "/exit",
+  NOT_FOUND: "*",
+};
